requirements.txt
# Tel-Insights Project Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Telegram Integration
telethon==1.34.0
python-telegram-bot==20.7

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Message Queue
pika==1.3.2

# AI/LLM Integration
google-generativeai==0.3.2
anthropic==0.7.8
openai==1.3.7

# Media Storage
google-cloud-storage==2.10.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
pillow==10.1.0

# FastMCP for Smart Analysis
fastmcp==0.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0

# Container Health Checks
healthcheck==1.3.3 